Move preprocesssing and stringification responsibility to renderer
