+++ packages/internals/renderer/src/renderer.js
@@ -1,3 +1,4 @@
+const toHTML = require('hast-util-to-html');
 const {Template} = require('@frctl/support');
 const EngineStore = require('./engine-store');
 
@@ -9,15 +10,29 @@ class Renderer {
     _engineStore.set(this, new EngineStore(engines));
   }
 
+  async preprocess(tpl, context = {}, opts = {}){
+    const props = _props.get(this);
+    if (isFunction(props.preprocess)) {
+      tpl = props.preprocess.bind(this)(tpl, context, opts);
+    }
+    return tpl;
+  }
+
   async render(template, context = {}, opts = {}) {
     if (!Template.isTemplate(template)) {
       throw new Error(`Renderer.render - template must be a template instance [template-invalid]`);
     }
+    const tpl = template.clone();
     const engine = _engineStore.get(this).getEngineFor(template.filename);
     if (!engine) {
       throw new Error(`No render engine found for template '${template.filename}' [engine-not-found]`);
     }
-    return engine.render(template.clone(), context, opts);
+    opts = Object.assign({}, opts, {template: tpl});
+    for (const preprocess of engine.preprocessors) {
+      preprocess(tpl.tree, context, opts);
+    }
+    const stringTemplate = toHTML(tpl.tree, opts.compiler);
+    return engine.render(stringTemplate, context, opts);
   }
 
 }
